# Fix _NoneLinearLSQFitter having a required extra arg