This should address https://github.com/astropy/astropy/pull/16673#issuecomment-2339334721 but I had to sacrifice the performance of `LevMarLSQFitter` to the gods of shitty code and backwards compatibility.

If this exercise has taught me anything, it's that `fitting.py` needs a major (breaking) refactor :fire:.

- [x] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.

closes #16979