@pllim 2024-09-10
How much sacrifice are we talking about here?

@mcara , are you satisfied with this fix?