@pllim 2024-09-10
For once, the benchmark results might be related.

```
| Change   | Before [7c77a0f3]    | After [060118be]    |   Ratio | Benchmark (Parameter)                                   |
|----------|----------------------|---------------------|---------|---------------------------------------------------------|
| +        | 4.20±0.3ms           | 5.91±0.3ms          |    1.41 | modeling.fitting.time_combined_gauss_1d_LevMarLSQFitter |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```