@pllim 2024-09-11
I think we're waiting for Tom's comment at https://github.com/astropy/astropy/pull/16982#discussion_r1752484325 to be resolved.