@mcara 2024-09-27
> > @mcara , are you satisfied with this fix?
> 
> @pllim I think this solves the issue of backward compatibility.

Sigh... not really. My bad: I should have run full tests. I didn't notice the change of `fitparam_indices`. This breaks old code.