@mcara 2024-09-27
https://github.com/spacetelescope/tweakwcs/actions/runs/11073910939/job/30771524455?pr=212#step:10:193

```
=========================== short test summary info ============================
FAILED tweakwcs/tests/test_linearfit.py::test_levmar2x2_multivariate - TypeError: _LevMarLSQFitter2x2.objective_function() got an unexpected keyword argument 'fit_param_indices'
```