@pllim 2024-09-27
Thanks for reporting back, @mcara . I opened https://github.com/astropy/astropy/issues/17085 follow-up issue that blocks v7.0 release until it is resolved.